Move repeated static messages to Messages.java
