+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -42,7 +42,7 @@ public class AddCommandIntegrationTest {
     public void execute_duplicatePatient_throwsCommandException() {
         Patient patientInList = model.getAddressBook().getPatientList().get(0);
         assertCommandFailure(new AddCommand(patientInList), model,
-                AddCommand.MESSAGE_DUPLICATE_PATIENT);
+                Messages.MESSAGE_DUPLICATE_PATIENT);
     }
 
 }
